+++ ChangeLog.md
@@ -3,9 +3,10 @@
 **Fork:**
 
 * (29 January 2015). Implemented Singleton pattern for `Replace_Plugin_Name_Common`, `Replace_Plugin_Name_Admin`, and `Replace_Plugin_Name_Public` via an abstract class.
+* (29 January 2015). Moved hooking functionality from a separate Loader class to all classes’ construct functions.
 * (3 January 2015). Added `$partials_path` properties for easier `@include` of HTML partials.
 * (3 January 2015). Adopt the Singleton design pattern for the main plugin class.
-* (3 January 2015). Added `Plugin_Name_Common` class for functionality that affect both public and admin areas of the site.
+* (3 January 2015). Added `Replace_Plugin_Name_Common` class for functionality that affect both public and admin areas of the site.
 * (3 January 2015). Find-and-replace support for all Plugin_Name etc. strings.
 
 **Original:**